✨ Added the ability to search for a character
This addition uses the Character pixy mixin which displays the characters with a size that may be too big compared to the other type search results.
As discussed on Discord this display can be adjusted in the future.

This addition also displays a maximum of 5 characters which is obviously open to being changed since it's a number picked just for testing and not for usability.

Lastly on the search.pixy file I left the comment below the addition since removing it or rewriting the addition without it broke my build.